// frontend/src/app/shared/interfaces/customer.ts
export interface Customer {
    customerId?: number;
    name?: string;
    nameKana?: string;
    nickname?: string;
    company?: string;
    client?: string;
    anniversaryOfCompany?: string;
    phone?: string;
    email?: string;
    lineId?: string;
    girlfriend?: string;
    birthday?: string;
    birthdayPresent?: string;
    summerGift?: string;
    yearEndGift?: string;
    birthPlace?: string;
    feature?: string;
    size?: string;
    isSmoke?: boolean;
    tabacco?: string;
    doGolf?: boolean;
    admirablePoint?: string;
    attractingPoint?: string;
    doDrink?: boolean;
    favoriteDrink?: string;
    unfavoriteDrink?: string;
    favoriteColor?: string;
    unfavoriteColor?: string;
    favoriteSports?: string;
    unfavoriteSports?: string;
    favoriteWoman?: string;
    unfavoriteWoman?: string;
    favoriteBrand?: string;
    unfavoriteBrand?: string;
    commonThings?: string;
    family?: string;
    pet?: string;
    address?: string;
    note?: string;
}